requirements.txt
-i https://pypi.org/simple
bcrypt==4.1.2; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
flask==3.0.3; python_version >= '3.8'
flask-dotenv==0.1.2
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.3; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
psycopg2==2.9.9; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
werkzeug==3.0.2; python_version >= '3.8'
